+++ crates/contracts/foundry/test/Costs.t.sol
@@ -29,8 +29,8 @@ contract CostsTest is KailuaTest {
     function test_propose_n(uint256 n) internal {
         // Deploy dispute contracts
         (treasury, game, anchor) = deployKailua(
-            uint256(4096 * n + 1), // 1 blob
-            uint256(0x01), // 1 block per commitment
+            uint64(4096 * n + 1), // 1 blob
+            uint64(0x01), // 1 block per commitment
             sha256(abi.encodePacked(bytes32(0x00))), // arbitrary genesis hash
             uint64(0x0), // genesis
             uint256(0), // start l2 from a while ago
